Refactors all event messages to have a common superclass.
