+++ src/org/openlcb/IdentifyProducersMessage.java
@@ -12,16 +12,11 @@ import edu.umd.cs.findbugs.annotations.*;
  */
 @Immutable
 @ThreadSafe
-public class IdentifyProducersMessage extends Message {
-    
+public class IdentifyProducersMessage extends EventMessage {
     public IdentifyProducersMessage(NodeID source, EventID eventID) {
-        super(source);
-        this.eventID = eventID;
+        super(source, eventID);
     }
         
-    EventID eventID;
-    public EventID getEventID() { return eventID; }
-    
     /**
      * Implement message-type-specific
      * processing when this message
@@ -32,19 +27,6 @@ public class IdentifyProducersMessage extends Message {
      @Override
      public void applyTo(MessageDecoder decoder, Connection sender) {
         decoder.handleIdentifyProducers(this, sender);
-     }
-
-     /**
-      * To be equal, messages have to have the
-      * same type and content
-      */
-     @Override
-     public boolean equals(Object o) {
-        if (!super.equals(o)) return false; // also checks type
-        IdentifyProducersMessage msg = (IdentifyProducersMessage) o;
-        if (! this.eventID.equals(msg.eventID))
-            return false;
-        return true;
      }
 
     public String toString() {
